requirements.txt
Pillow==10.0.0
Pillow==10.2.0
websocket_client==1.7.0
